split priorities into proximities and priorities
